Trapping Exceptions for when the solidfire cluster is in unhealthy state.
While performing a cluster update, we saw that some API calls fail when the cluster is not in completely healthy state. This makes the ElementFactory Connection to timeout and hence trigger an exception which was not trapped.

The patch changes the logic to open a new connection everytime we beging doing a metrics polling, that's why I also modified the timeout to 15seconds.